ENH: Fix loading bluebrints when --team-name is specified; STYLE: Use default-member-initializer instead of assigning default values in constructor

+++ Modules/Components/itkImageRegistrationMethodv4/include/selxItkImageRegistrationMethodv4Component.h
@@ -148,7 +148,7 @@ private:
 
   ComponentBase::ParameterValueType m_RescaleIntensity;
   bool m_InvertIntensity;
-  float m_MetricSamplingPercentage;
+  double m_MetricSamplingPercentage;
 
 protected:
